Fixed issue with tax calculation

+++ Course/Services/ITaxService.cs
@@ -2,7 +2,7 @@
 {
     interface ITaxService
     {
-        double IcmsValue(double amount, int icmsAliquot);
-        double IpiValue(double amount, int ipiAliquot);
+        double IcmsValue(double amount, double icmsAliquot);
+        double IpiValue(double amount, double ipiAliquot);
     }
 }
